cyclical dependence
