+++ proxy/src/main/java/com/jhl/web/webConf/WebConfig.java
@@ -8,6 +8,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.MediaType;
 import org.springframework.http.converter.HttpMessageConverter;
+import org.springframework.web.client.RestTemplate;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
@@ -40,5 +41,17 @@ public class WebConfig implements WebMvcConfigurer {
         FastJsonHttpMessageConverter fastJsonHttpMessageConverter = new FastJsonHttpMessageConverter();
         fastJsonHttpMessageConverter.setSupportedMediaTypes(Lists.newArrayList(MediaType.APPLICATION_JSON));
         return fastJsonHttpMessageConverter;
+    }
+
+    @Bean
+    public RestTemplate getRestTemplate() {
+        RestTemplate restTemplate = new RestTemplate();
+        restTemplate.getMessageConverters().removeIf(httpMessageConverter -> {
+            String name = httpMessageConverter.getClass().getName();
+            if (name.contains("json")) return true;
+            else return false;
+        });
+        restTemplate.getMessageConverters().add(fastJsonHttpMessageConverter());
+        return restTemplate;
     }
 }
