+++ vpn-admin/src/main/java/com/jhl/admin/config/WebConfig.java
@@ -3,11 +3,15 @@ package com.jhl.admin.config;
 import com.alibaba.fastjson.support.spring.FastJsonHttpMessageConverter;
 import com.google.common.collect.Lists;
 import com.jhl.admin.Interceptor.AuthInterceptor;
+import com.jhl.admin.constant.ProxyConstant;
+import org.apache.commons.codec.digest.DigestUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.http.converter.HttpMessageConverter;
+import org.springframework.web.client.RestTemplate;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
@@ -19,12 +23,11 @@ import java.util.List;
 public class WebConfig implements WebMvcConfigurer {
     @Autowired
     AuthInterceptor authInterceptor;
-
     @Autowired
-    FastJsonHttpMessageConverter fastJsonHttpMessageConverter;
+    ProxyConstant proxyConstant;
     @Override
     public void extendMessageConverters(List<HttpMessageConverter<?>> converters) {
-        converters.add(0, fastJsonHttpMessageConverter);
+        converters.add(0, fastJsonHttpMessageConverter());
 
     }
 
@@ -35,6 +38,19 @@ public class WebConfig implements WebMvcConfigurer {
         return fastJsonHttpMessageConverter;
     }
 
+
+    @Bean
+    public RestTemplate restTemplate() {
+        RestTemplate restTemplate = new RestTemplate();
+        restTemplate.getMessageConverters().add(fastJsonHttpMessageConverter());
+        restTemplate.getInterceptors().add((request, body, execution) -> {
+            HttpHeaders headers = request.getHeaders();
+            headers.add("Authorization", DigestUtils.md5Hex(proxyConstant.getAuthPassword()));
+            headers.setContentType(MediaType.APPLICATION_JSON);
+            return execution.execute(request, body);
+        });
+        return restTemplate;
+    }
     /**
      * 拦截器
      *
